refactor: Proper tasks for muzzle reports
