+++ buildSrc/src/main/kotlin/datadog/gradle/plugin/muzzle/tasks/AbstractMuzzleReportTask.kt
@@ -1,0 +1,32 @@
+package datadog.gradle.plugin.muzzle.tasks
+
+import datadog.gradle.plugin.muzzle.TestedArtifact
+import java.util.SortedMap
+
+abstract class AbstractMuzzleReportTask : AbstractMuzzleTask() {
+  internal fun dumpVersionsToCsv(versions: SortedMap<String, TestedArtifact>) {
+    val filename = project.path.replaceFirst("^:".toRegex(), "").replace(":", "_")
+
+    val versionsFile = project.rootProject.layout.buildDirectory.file("${MUZZLE_DEPS_RESULTS}/$filename.csv")
+    with(project.file(versionsFile)) {
+      parentFile.mkdirs()
+      writeText("instrumentation,jarGroupId,jarArtifactId,lowestVersion,highestVersion\n")
+      versions.values.forEach {
+        appendText(
+          listOf(
+            it.instrumentation,
+            it.group,
+            it.module,
+            it.lowVersion.toString(),
+            it.highVersion.toString()
+          ).joinToString(",") + "\n"
+        )
+      }
+      project.logger.info("Wrote muzzle versions report to\n  $this")
+    }
+  }
+
+  companion object {
+    internal const val MUZZLE_DEPS_RESULTS = "muzzle-deps-results"
+  }
+}
